+++ internal/ent/namebhl/interface.go
@@ -1,8 +1,9 @@
 package namebhl
 
+import "github.com/bits-and-blooms/bloom/v3"
+
 // NameBHL interface provides methods to collect data from BHLindex.
 type NameBHL interface {
-	PageFilesToIDs() error
-	ImportNames() error
-	ImportOccurrences() error
+	ImportNames() (*bloom.BloomFilter, error)
+	ImportOccurrences(*bloom.BloomFilter) error
 }