+++ internal/io/namesbhlio/db-namesbhlio.go
@@ -6,6 +6,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/bits-and-blooms/bloom/v3"
 	"github.com/dustin/go-humanize"
 	"github.com/lib/pq"
 )
@@ -17,30 +18,31 @@ const (
 	OccurrencesNumberF   = 3
 	OddsLog10F           = 4
 	MatchTypeF           = 5
-	EditDistanceF        = 6
-	StemEditDistanceF    = 7
-	MatchedCanonicalF    = 8
-	MatchedFullNameF     = 9
-	MatchedCardinalityF  = 10
-	CurrentCanonicalF    = 11
-	CurrentFullNameF     = 12
-	CurrentCardinalityF  = 13
-	ClassificationF      = 14
-	ClassificationRanksF = 15
-	ClassificationIDsF   = 16
-	RecordIDF            = 17
-	DataSourceIDF        = 18
-	DataSourceF          = 19
-	DataSourcesNumberF   = 20
-	CurationF            = 21
-	ErrorF               = 22
+	MatchSortOrderF      = 6
+	EditDistanceF        = 7
+	StemEditDistanceF    = 8
+	MatchedCanonicalF    = 9
+	MatchedFullNameF     = 10
+	MatchedCardinalityF  = 11
+	CurrentCanonicalF    = 12
+	CurrentFullNameF     = 13
+	CurrentCardinalityF  = 14
+	ClassificationF      = 15
+	ClassificationRanksF = 16
+	ClassificationIDsF   = 17
+	RecordIDF            = 18
+	DataSourceIDF        = 19
+	DataSourceF          = 20
+	DataSourcesNumberF   = 21
+	CurationF            = 22
+	ErrorF               = 23
 )
 
-func (n namesbhlio) saveNames(ch <-chan [][]string) error {
+func (n namesbhlio) saveNames(ch <-chan [][]string, blf *bloom.BloomFilter) error {
 	total := 0
 	columns := []string{"id", "name", "record_id", "match_type",
-		"edit_distance", "stem_edit_distance", "matched_name", "matched_canonical",
-		"current_name", "current_canonical", "classification",
+		"match_sort_order", "edit_distance", "stem_edit_distance", "matched_name",
+		"matched_canonical", "current_name", "current_canonical", "classification",
 		"classification_ranks", "classification_ids", "data_source_id",
 		"data_source_title", "data_sources_number", "curation", "occurences",
 		"odds_log10", "error"}
@@ -56,7 +58,7 @@ func (n namesbhlio) saveNames(ch <-chan [][]string) error {
 			return err
 		}
 
-		var eDist, stemDist, dsID, dsNum, occurs int
+		var eDist, stemDist, dsID, matchSort, dsNum, occurs int
 		var odds float64
 
 		for _, v := range names {
@@ -68,6 +70,9 @@ func (n namesbhlio) saveNames(ch <-chan [][]string) error {
 				dsID, err = strconv.Atoi(v[DataSourceIDF])
 			}
 			if err == nil {
+				matchSort, err = strconv.Atoi(v[MatchSortOrderF])
+			}
+			if err == nil {
 				dsNum, err = strconv.Atoi(v[DataSourcesNumberF])
 			}
 			if err == nil {
@@ -77,13 +82,20 @@ func (n namesbhlio) saveNames(ch <-chan [][]string) error {
 				return err
 			}
 
+			// for now only take Catalogue of Life names.
+			if dsID != 1 {
+				continue
+			}
+
+			blf.Add([]byte(v[NameIDF]))
+
 			odds, err = strconv.ParseFloat(v[OddsLog10F], 64)
 			if err != nil {
 				odds = 0
 			}
 
 			_, err = stmt.Exec(v[NameIDF], v[DetectedNameF], v[RecordIDF],
-				v[MatchTypeF], eDist, stemDist, v[MatchedFullNameF],
+				v[MatchTypeF], matchSort, eDist, stemDist, v[MatchedFullNameF],
 				v[MatchedCanonicalF], v[CurrentFullNameF], v[CurrentCanonicalF],
 				v[ClassificationF], v[ClassificationRanksF], v[ClassificationIDsF],
 				dsID, v[DataSourceF], dsNum, true, occurs, odds, v[ErrorF])